BOOST_PP_TUPLE_POP_FRONT() added & files reorganized
