+++ include/sqlpp11/ppgen/colops/blob.h
@@ -24,12 +24,27 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef _sqlpp__ppgen__engine_h
-#define _sqlpp__ppgen__engine_h
+#ifndef _sqlpp__ppgen__colops__blob_h
+#define _sqlpp__ppgen__colops__blob_h
 
-#define SQLPP_DECLARE_TABLE_GET_PROC_LAZY_SQLPP_ENGINE(str) \
-	PROC_SQLPP_ENGINE
-#define SQLPP_DECLARE_TABLE_GEN_PROC_SQLPP_ENGINE(str) \
-	[ENGINE is not implemented]
+#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_tinyblob \
+	PROC_tinyblob
+#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_tinyblob(...) \
+	sqlpp::blob
 
-#endif // _sqlpp__ppgen__engine_h
+#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_blob \
+	PROC_blob
+#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_blob(...) \
+	sqlpp::blob
+
+#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_mediumblob \
+	PROC_mediumblob
+#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_mediumblob(...) \
+	sqlpp::blob
+
+#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_longblob \
+	PROC_longblob
+#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_longblob(...) \
+	sqlpp::blob
+
+#endif // _sqlpp__ppgen__colops__blob_h
