+++ include/sqlpp11/ppgen/colops/bool.h
@@ -24,12 +24,12 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef _sqlpp__ppgen__text_h
-#define _sqlpp__ppgen__text_h
+#ifndef _sqlpp__ppgen__colops__bool_h
+#define _sqlpp__ppgen__colops__bool_h
 
-#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_text \
-	PROC_text
-#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_text(...) \
-	sqlpp::text
+#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_bool \
+	PROC_bool
+#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_bool(...) \
+	sqlpp::boolean
 
-#endif // _sqlpp__ppgen__text_h
+#endif // _sqlpp__ppgen__colops__bool_h
