+++ include/sqlpp11/ppgen/colops/null.h
@@ -24,27 +24,12 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef _sqlpp__ppgen__blob_h
-#define _sqlpp__ppgen__blob_h
+#ifndef _sqlpp__ppgen__colops__null_h
+#define _sqlpp__ppgen__colops__null_h
 
-#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_tinyblob \
-	PROC_tinyblob
-#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_tinyblob(...) \
-	sqlpp::blob
+#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_SQLPP_NULL \
+	PROC_SQLPP_NULL
+#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_SQLPP_NULL(...) \
+	sqlpp::tag::can_be_null
 
-#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_blob \
-	PROC_blob
-#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_blob(...) \
-	sqlpp::blob
-
-#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_mediumblob \
-	PROC_mediumblob
-#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_mediumblob(...) \
-	sqlpp::blob
-
-#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_longblob \
-	PROC_longblob
-#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_longblob(...) \
-	sqlpp::blob
-
-#endif // _sqlpp__ppgen__blob_h
+#endif // _sqlpp__ppgen__colops__null_h
