+++ include/sqlpp11/ppgen/colops/primary_key.h
@@ -24,27 +24,12 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef _sqlpp__ppgen__integer_h
-#define _sqlpp__ppgen__integer_h
+#ifndef _sqlpp__ppgen__colops__primary_key_h
+#define _sqlpp__ppgen__colops__primary_key_h
 
-#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_tinyint \
-	PROC_tinyint
-#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_tinyint(...) \
-	sqlpp::tinyint
+#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_SQLPP_PRIMARY_KEY \
+	PROC_SQLPP_PRIMARY_KEY
+#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_SQLPP_PRIMARY_KEY(...) \
+	[PRIMARY_KEY is not implemented]
 
-#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_smallint \
-	PROC_smallint
-#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_smallint(...) \
-	sqlpp::smallint
-
-#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_int \
-	PROC_int
-#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_int(...) \
-	sqlpp::integer
-
-#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_bigint \
-	PROC_bigint
-#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_bigint(...) \
-	sqlpp::bigint
-
-#endif // _sqlpp__ppgen__integer_h
+#endif // _sqlpp__ppgen__colops__primary_key_h
