+++ include/sqlpp11/ppgen/tblops/comment.h
@@ -24,12 +24,12 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef _sqlpp__ppgen__timestamp_h
-#define _sqlpp__ppgen__timestamp_h
+#ifndef _sqlpp__ppgen__tblops__comment_h
+#define _sqlpp__ppgen__tblops__comment_h
 
-#define SQLPP_DECLARE_COLUMN_GET_TRAITS_LAZY_timestamp \
-	PROC_timestamp
-#define SQLPP_DECLARE_COLUMN_GEN_TRAITS_PROC_timestamp(...) \
-	[timestamp is not implemented]
+#define SQLPP_DECLARE_TABLE_GET_TRAITS_LAZY_SQLPP_COMMENT(str) \
+	PROC_SQLPP_COMMENT
+#define SQLPP_DECLARE_TABLE_GEN_TRAITS_PROC_SQLPP_COMMENT(str) \
+	[COMMENT is not implemented]
 
-#endif // _sqlpp__ppgen__timestamp_h
+#endif // _sqlpp__ppgen__tblops__comment_h
